fix(firmware): clear the clang-tidy baseline, and analyse the ARM build
The CI job ran clang-tidy over all 67 firmware sources using the *SITL*
compile database. Eleven of them are not in that database, so they were
analysed with default flags and reported undeclared-function and parse
errors that do not exist in the flight build.

Split it: one job over the ARM database (all 67 sources, via the new
tools/dev/run_clang_tidy.sh, whose --extra-args are each documented with
the false-positive class they remove -- without the newlib -isystem, five
files die on "'string.h' file not found" and are not analysed at all), and
one over the SITL database restricted to the files that database actually
contains. Both configurations are now at zero findings.

Two of the findings were real defects, not noise:

  * navlink_tx.c called angle_controller_height_state() with no prototype
    in scope -- the one new warning in an otherwise warning-clean flight
    build. Correct code by luck of the AAPCS, a constraint violation since
    C99, an error by default in GCC 14.
  * vertical_task.c reached vayu_log transitively.

The rest are widening fixes on index arithmetic, plus two places where the
check is answered with a comment rather than obeyed: pid_config.c (int8_t
holding +/-1, where the suggested unsigned round-trip would read -1 as 255)
and bmx160.c, which now documents that bmx160_acc_range_t's enumerators are
g-values while the fields hold register codes -- the trap that will make the
first attempt at configuring the IMU look like it worked.

attitude_task.c now derives the reported estimator rate from the decimation
instead of reporting the nominal target, which stops agreeing with reality
as soon as the sample rate does not divide evenly.

+++ firmware/src/comm/ibus.c
@@ -59,7 +59,8 @@ bool ibus_parse_byte(uint8_t b, ibus_data_t *data) {
       // Packet valid!
       if (data) {
         for (int i = 0; i < IBUS_MAX_CHANNELS; i++) {
-          data->channels[i] = buffer[2 + i * 2] | (buffer[3 + i * 2] << 8);
+          data->channels[i] =
+              (uint16_t)(buffer[2 + i * 2] | (buffer[3 + i * 2] << 8));
         }
         /* FlySky has no in-protocol failsafe flag — link-loss is inferred from
          * the throttle channel (rc_throttle_failsafe_step) and the staleness
